% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cell2go}
\alias{cell2go}
\title{Get GO ids from cell types}
\usage{
cell2go(celltype_qids)
}
\arguments{
\item{celltype_qids}{A character vector of cell type Wikidata IDs}
}
\value{
A dataframe of 5 columns: The first two correspond to cell types,
   that is, their Wikidata Identifier and their name. The two following
   columns correspond to the Gene Ontology IDs given, and the last column
   corresponds to the cell type's marker that led to the result.
}
\description{
Return GO identifiers related to gene markers
of specific cell types.
}
\details{
This function takes in a character vector of
cell type Q identifiers (QIDs), queries Wikidata
(\url{https://www.wikidata.org/})
and a local RDF database to return GO IDs related
to gene markers of these cell types.
}
\examples{
# Return GO IDs related to cell type
# 'human smooth muscle cell'
cell2go('Q101404901')
\dontrun{
# IDs should always start with 'Q'
go2cell('101404901')
}
}
\seealso{
\url{https://www.wikidata.org/} for information regarding
  Wikidata and \url{https://panglaodb.se/index.html} for the
  original database this information was adapted from. For an
  example of a cell type item in Wikidata, see
  \url{https://www.wikidata.org/wiki/Q101405098}
}
