% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype_marker_turtle}
\alias{celltype_marker_turtle}
\title{RDF graph of cell type - marker correspondence}
\format{
A character vector of a Turtle RDF graph
}
\source{
\url{https://panglaodb.se/}
}
\usage{
celltype_marker_turtle
}
\description{
A dataset consisting of a Turtle-syntax RDF graph
of human cell types and their corresponding markers,
extracted, manually curated and serialized from
\url{https://panglaodb.se/}
}
\keyword{datasets}
