% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.collapse_as_values}
\alias{.collapse_as_values}
\title{Format string or character vector as SPARQL-compatible list}
\usage{
.collapse_as_values(item_vector, quotes = FALSE)
}
\arguments{
\item{item_vector}{A string or a character vector of items}

\item{quotes}{A logical to indicate whether or not to encase in quotes}
}
\value{
A SPARQL-compatible string of the values
}
\description{
Format string or character vector as SPARQL-compatible list
}
\examples{
.collapse_as_values(c("GO:0006936", "GO:0008152"))
}
