% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.get_celltype_items}
\alias{.get_celltype_items}
\title{Query Wikidata for cell type item names from cell type QIDs}
\usage{
.get_celltype_items(celltype_ids)
}
\arguments{
\item{celltype_ids}{A character vector of cell type QIDs}
}
\value{
A dataframe of QIDs and their respective labels
}
\description{
Query Wikidata for cell type item names from cell type QIDs
}
\examples{
.get_celltype_items('(wd:Q101405098) (wd:Q101404862)')
}
