% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.get_genes_from_go}
\alias{.get_genes_from_go}
\title{Query Wikidata for genes related to GO ids}
\usage{
.get_genes_from_go(go_ids)
}
\arguments{
\item{go_ids}{A character vector of GO identifiers}
}
\value{
A dataframe of GO ids and genes
}
\description{
Query Wikidata for genes related to GO ids
}
\examples{
.get_genes_from_go("('GO:0006936')")
}
