% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.get_go_from_genes}
\alias{.get_go_from_genes}
\title{Query Wikidata for GO ids associated with particular genes}
\usage{
.get_go_from_genes(genes)
}
\arguments{
\item{genes}{A character vector of SPARQL-valid gene items}
}
\value{
A dataframe of GO ids and genes
}
\description{
Query Wikidata for GO ids associated with particular genes
}
\examples{
.get_go_from_genes("(wd:Q15317282) (wd:Q14912592) (wd:Q14912176)")
}
