% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_species_item}
\alias{.get_species_item}
\title{Return the item for a particular species}
\usage{
.get_species_item(species_name)
}
\arguments{
\item{species_name}{The scientific name of the species}
}
\value{
The Wikidata item for said species
}
\description{
Return the item for a particular species
}
