% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.remove_wdt_url}
\alias{.remove_wdt_url}
\title{Substitute the wikidata url for a SPARQL-friendly wd:}
\usage{
.remove_wdt_url(column)
}
\arguments{
\item{column}{The column of the dataframe to substitute}
}
\value{
A character vector
}
\description{
Substitute the wikidata url for a SPARQL-friendly wd:
}
