% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{go2cell}
\alias{go2cell}
\title{Get cell types from GO ids}
\usage{
go2cell(go_ids, species = "Homo sapiens")
}
\arguments{
\item{go_ids}{A character vector of Gene Ontology identifiers.}

\item{species}{A character of a species, can either be Homo sapiens
(default) or Mus musculus.}
}
\value{
A dataframe of 5 columns: The first two correspond to cell types,
   that is, their Wikidata Identifier and their name. The two following
   columns correspond to the Gene Ontology IDs given, and the last column
   corresponds to the cell type's marker that led to the result.
}
\description{
Return cell type information for specific Gene Ontology identifiers
}
\details{
This function takes in a character vector of
Gene Ontology (GO) identifiers, queries Wikidata
(\url{https://www.wikidata.org/})
for genes related to said identifiers, and crosses this data
with a local RDF database of cell type-marker correspondence,
returning, at last, a dataframe of cell type information related to
the given GO identifiers.
}
\examples{
# Return cell types related to 'muscle contraction' and 'metabolism'
go2cell(c("GO:0006936", "GO:0008152"))

# Return cell types related to 'response to zinc ion'
go2cell("GO:0010043")
\dontrun{
# IDs should use ':' not '_'
go2cell("GO_0010043")
}
}
\seealso{
\url{https://www.wikidata.org/} for information regarding
  Wikidata and \url{https://panglaodb.se/index.html} for the
  original database this information was adapted from.
}
