% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go2cell_test_data}
\alias{go2cell_test_data}
\title{A dataframe of Gene symbols and GO IDs}
\format{
A dataframe of 3 columns, corresponding to
   a gene name, a go ID process term and the go ID
   itself.
}
\usage{
go2cell_test_data
}
\description{
Small dataset for testing and showcase purposes.
}
\author{
João Vitor F Cavalcante
}
\keyword{datasets}
