% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.query_ctp_turtle}
\alias{.query_ctp_turtle}
\title{Query local database for cell types related to particular markers}
\usage{
.query_ctp_turtle(gene_items)
}
\arguments{
\item{gene_items}{A character vector of markers}
}
\value{
A dataframe of cell type Wikidata items and their respective markers
}
\description{
Query local database for cell types related to particular markers
}
\examples{
.query_ctp_turtle("(wd:Q14881255) (wd:Q18016342)")
}
