% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.query_ctp_turtle}
\alias{.query_ctp_turtle}
\title{Query local database for cell types related to particular markers}
\usage{
.query_ctp_turtle(query_values, query_key, selector)
}
\arguments{
\item{query_values}{A character vector of SPARQL-valid values}

\item{query_key}{A character of the key value to select items by}

\item{selector}{A character vector of the query_values entity}
}
\value{
A dataframe of cell type Wikidata items and their respective markers
}
\description{
Query local database for cell types related to particular markers
}
\examples{
.query_ctp_turtle("(wd:Q14881255) (wd:Q18016342)")
}
